CODE/NL2SQL_DAX_UNIFIED_PIPELINE/requirements.txt
# Core dependencies for NL2SQL&DAX Unified Pipeline
# ===================================================

# Database connectivity
pyodbc>=4.0.39
sqlalchemy>=1.4.0

# Azure services
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# OpenAI and AI services
openai>=1.12.0
tiktoken>=0.5.2

# Web framework
streamlit>=1.32.0
streamlit-option-menu>=0.3.6

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# HTTP requests
requests>=2.31.0
urllib3>=2.0.0

# Logging and monitoring
structlog>=23.2.0

# Date and time handling
python-dateutil>=2.8.2

# JSON handling
jsonschema>=4.20.0

# File operations
pathlib2>=2.3.7

# Testing (development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Code quality (development)
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0

# Documentation (development)
mkdocs>=1.5.0
mkdocs-material>=9.5.0